Fixed stuff
