+++ src/main/java/com/gmail/goosius/siegewar/listeners/SiegeWarTownEventListener.java
@@ -18,7 +18,7 @@ import com.gmail.goosius.siegewar.settings.SiegeWarSettings;
 import com.gmail.goosius.siegewar.settings.Translation;
 import com.gmail.goosius.siegewar.utils.SiegeWarDistanceUtil;
 import com.gmail.goosius.siegewar.utils.SiegeWarPermissionUtil;
-import com.gmail.goosius.siegewar.utils.TownPeacefulnessUtil;
+import com.gmail.goosius.siegewar.utils.SiegeWarTownUtil;
 import com.palmergames.bukkit.towny.TownyFormatter;
 import com.palmergames.bukkit.towny.TownyMessaging;
 import com.palmergames.bukkit.towny.TownySettings;
@@ -156,7 +156,7 @@ public class SiegeWarTownEventListener implements Listener {
 			TownMetaDataController.setDesiredPeacefulnessSetting(town, event.getFutureState());
 			TownMetaDataController.setPeacefulnessChangeDays(town, 0);
 			if (event.getFutureState() == true)
-				TownPeacefulnessUtil.disableTownPVP(town);
+				SiegeWarTownUtil.disableTownPVP(town);
 			return;
 		} else {
 			int days;
